requirements.txt
pillow
robotframework >= 3.1.1
robotframework-seleniumlibrary >= 3.3.1
robotframework-sshlibrary
robotframework-requests
webdrivermanager